alphabet/requirements.txt
backcall==0.1.0
certifi==2019.11.28
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
decorator==4.4.1
Django==2.2.7
django-dynamic-db-router==0.3.0
django-rest-swagger==2.2.0
djangorestframework==3.10.3
idna==2.8
ipython==7.9.0
ipython-genutils==0.2.0
itypes==1.1.0
jedi==0.15.1
Jinja2==2.10.3
MarkupSafe==1.1.1
mysqlclient==1.4.6
openapi-codec==1.3.2
parso==0.5.1
pexpect==4.7.0
pickleshare==0.7.5
prompt-toolkit==2.0.10
ptyprocess==0.6.0
Pygments==2.4.2
pytz==2019.3
requests==2.22.0
simplejson==3.17.0
six==1.13.0
sqlparse==0.3.0
traitlets==4.3.3
uritemplate==3.0.0
urllib3==1.25.7
wcwidth==0.1.7
